[refactor] New EmbeddedXMLStreamReader which can read beyond an element boundary to see siblings
